Update nathelper.c
Function does not proceed beyond call to "alter_mediaip()" (at least this is the case in Fedora) due to the multi-line comment in the function call and therefore is not returning.

This can be shown by placing an LM_INFO() after the call and executing fix_nated_sdp("8","$si"). The error is not evident because the first iteration on modifying the o= or c= is successful, but the code intends to modify all o/c= in the body.